Tweaked animations on LinkTab and implemented underline under MainNav via current page
