--- components/profile/Profile.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Profile = () => {
-  return (
-    <div className="WRAPPER flex flex-col w-full h-full">This is profile</div>
-  );
-};
-
-export default Profile;
