+++ package.json
@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@headlessui/react": "1.6.6",
+    "@headlessui/react": "^1.6.6",
     "@heroicons/react": "1.0.6",
     "@next-auth/mongodb-adapter": "^1.0.3",
     "bcryptjs": "^2.4.3",
