+++ src/components/dashboard/components/LinkTab.js
@@ -2,33 +2,33 @@ import React, { useState, Fragment } from "react";
 import {
   PhotographIcon,
   TrashIcon,
-  ChevronUpIcon
+  ChevronUpIcon,
 } from "@heroicons/react/solid";
 import TitleInput from "./TitleInput";
 import URLInput from "./URLInput";
 
 import { Dialog, Transition, Disclosure } from "@headlessui/react";
 
-const DisclosureTab = ({ isOpen, setIsOpen, handleDeleteLink, _id }) => {
+const BottomBar = ({ isOpen, setIsOpen, handleDeleteLink, _id }) => {
   return (
     <Disclosure>
       {({ open, close }) => (
         <>
           <div className="flex flex-row w-full justify-between px-6">
             <PhotographIcon className="h-5 w-5 text-green-300 hover:text-green-400 cursor-pointer" />
-            <Disclosure.Button className="">
+            <Disclosure.Button onClick={() => setIsOpen(true)}>
               <TrashIcon className="h-5 w-5 text-slate-300 hover:text-slate-400 cursor-pointer" />
             </Disclosure.Button>
           </div>
           <Transition
             enter="transition duration-100 ease-out"
-            enterFrom="transform scale-95 opacity-0"
-            enterTo="transform scale-100 opacity-100"
-            leave="transition duration-75 ease-out"
-            leaveFrom="transform scale-100 opacity-100"
-            leaveTo="transform scale-95 opacity-0"
+            enterFrom="transform height-[0px] translate-y-[-20px] opacity-20"
+            enterTo="transform height-[100px] translate-y-[0px] opacity-100"
+            leave="transition duration-100 ease-out"
+            leaveFrom="transform translate-y-[0px] opacity-100"
+            leaveTo="transform translate-y-[-10px] opacity-0"
           >
-            {open && (
+            {isOpen && (
               <Disclosure.Panel className="flex flex-col pt-4 pb-2 text-sm text-gray-900 text-center gap-4">
                 <div className="DELETE HEAD w-full bg-slate-200 text-lg text-center font-semibold">
                   Delete
@@ -36,7 +36,10 @@ const DisclosureTab = ({ isOpen, setIsOpen, handleDeleteLink, _id }) => {
                 <span className="w-full">Delete forever?</span>
                 <div className="flex flex-row px-5 gap-2">
                   <button
-                    onClick={() => close()}
+                    onClick={() => {
+                      close();
+                      setIsOpen(false);
+                    }}
                     className="px-5 py-3 w-full bg-slate-200 rounded-[1000px]"
                   >
                     Cancel
@@ -57,8 +60,6 @@ const DisclosureTab = ({ isOpen, setIsOpen, handleDeleteLink, _id }) => {
   );
 };
 
-// handleDeleteLink.mutate(_id)
-
 const LinkTab = ({ _id, handleDeleteLink, handleUpdateLink, url, title }) => {
   const [isOpen, setIsOpen] = useState(false);
 
@@ -73,7 +74,9 @@ const LinkTab = ({ _id, handleDeleteLink, handleUpdateLink, url, title }) => {
 
   return (
     <>
-      <div className="flex flex-col w-full bg-white rounded-[33px] py-4 gap-3">
+      <div
+        className={`flex flex-col w-full h-auto bg-white rounded-[33px] py-4 gap-3 justify-between transition-all ease-in-out duration-200 `}
+      >
         <div className="flex flex-col py-1 px-6">
           <TitleInput
             className={"font-bold"}
@@ -84,7 +87,7 @@ const LinkTab = ({ _id, handleDeleteLink, handleUpdateLink, url, title }) => {
           />
           <URLInput data={url} label={"Url"} handleSubmit={handleSubmit} />
         </div>
-        <DisclosureTab
+        <BottomBar
           isOpen={isOpen}
           setIsOpen={setIsOpen}
           handleDeleteLink={handleDeleteLink}
