+++ src/pages/api/auth/register.js
