+++ src/pages/dashboard/index.js
