+++ src/pages/middleware.js
@@ -1,0 +1,6 @@
+export { default } from "next-auth/middleware";
+
+export const config = {
+  matcher: ["/dashboard"],
+  secret: process.env.NEXTAUTH_SECRET
+};
